pyproject.toml
[tool.poetry]
name = "fastapi-kafka"
version = "0.1.0"
description = ""
authors = ["Janina <janina.volkmer@elements.nl>"]
readme = "README.md"
packages = [{include = "fastapi_kafka"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
